# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GuestFlow API
  description: Backend API for the GuestFlow digital experience platform
  version: 1.0.0
  contact:
    name: Digital Experience Team
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/checkins:
    get:
      summary: List all check-ins
      operationId: getCheckins
      tags: [Check-ins]
      responses:
        '200':
          description: List of check-ins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CheckIn'
    post:
      summary: Create a guest check-in
      operationId: createCheckin
      tags: [Check-ins]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckInRequest'
      responses:
        '201':
          description: Check-in created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckIn'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Experience not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    CheckInRequest:
      type: object
      required: [guestName, experienceId, reservationCode]
      properties:
        guestName:
          type: string
          example: Ana Silva
        experienceId:
          type: string
          example: exp-001
        reservationCode:
          type: string
          example: GF-2048
    CheckIn:
      type: object
      properties:
        checkinId:
          type: string
          example: chk-1700000000000-abc12
        guestName:
          type: string
        experienceId:
          type: string
        experienceName:
          type: string
        reservationCode:
          type: string
        status:
          type: string
          example: CONFIRMED
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          type: string